messages_2/requirements.txt
flask
requests
hazelcast-python-client
consulate